aesthetics and avn-project date fix
There was an issue with the check on if closing a project preceded its deadline.

Invoke-AVNSignOn now doesn't add turns to current turns but sets it to the daily alottment.

Added Close-AVNSeason to help.

Removed warning about running the module from a non-production location. I can probably remove this line from the config file.